# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for Phase 1 user top-up campaigns (admin + user-facing). Operates under `/campaign-center-api/v1`;",
        "title": "Campaign Center API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/campaign-center-api/v1",
    "paths": {
        "/admin/campaigns": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign"
                ],
                "summary": "List campaigns (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Campaign status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campaign type e.g. TOPUP_REWARD",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign"
                ],
                "summary": "Create campaign (admin)",
                "parameters": [
                    {
                        "description": "Campaign payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateCampaignReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "validation error",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/campaigns/{campaignId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign"
                ],
                "summary": "Get campaign detail (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign"
                ],
                "summary": "Update campaign (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campaign payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateCampaignReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "409": {
                        "description": "not draft",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/campaigns/{campaignId}/archive": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign"
                ],
                "summary": "Archive campaign (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Operator",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PublishOperatorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "409": {
                        "description": "not eligible or already archived",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/campaigns/{campaignId}/publish": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign"
                ],
                "summary": "Publish campaign (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Operator",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PublishOperatorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.CreateCampaignReq": {
            "type": "object",
            "required": [
                "campaignEndTime",
                "campaignStartTime",
                "name",
                "registrationEndTime",
                "registrationStartTime",
                "rewardRules",
                "targetMarket",
                "targetUserSegment",
                "type"
            ],
            "properties": {
                "campaignEndTime": {
                    "type": "string"
                },
                "campaignStartTime": {
                    "type": "string"
                },
                "landingPageId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "registrationEndTime": {
                    "type": "string"
                },
                "registrationStartTime": {
                    "type": "string"
                },
                "rewardRules": {
                    "$ref": "#/definitions/api.RewardRulesReq"
                },
                "targetMarket": {
                    "type": "string"
                },
                "targetUserSegment": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.PublishOperatorReq": {
            "type": "object",
            "required": [
                "operator"
            ],
            "properties": {
                "operator": {
                    "type": "string"
                }
            }
        },
...
        "api.RewardRulesReq": {
            "type": "object",
            "required": [
                "maxClaimPerUser",
                "rewardType",
                "topupThreshold"
            ],
            "properties": {
                "maxClaimPerUser": {
                    "type": "integer"
                },
                "maxRewardAmount": {
                    "type": "number"
                },
                "minObtainDays": {
                    "type": "integer"
                },
                "rewardAmount": {
                    "type": "number"
                },
                "rewardCurrency": {
                    "type": "string"
                },
                "rewardMode": {
                    "type": "string"
                },
                "rewardPercentage": {
                    "type": "number"
                },
                "rewardType": {
                    "type": "string"
                },
                "topupThreshold": {
                    "type": "number"
                }
            }
        },
...
        "api.UpdateCampaignReq": {
            "type": "object",
            "required": [
                "campaignEndTime",
                "campaignStartTime",
                "name",
                "registrationEndTime",
                "registrationStartTime",
                "rewardRules",
                "targetMarket",
                "targetUserSegment"
            ],
            "properties": {
                "campaignEndTime": {
                    "type": "string"
                },
                "campaignStartTime": {
                    "type": "string"
                },
                "landingPageId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "registrationEndTime": {
                    "type": "string"
                },
                "registrationStartTime": {
                    "type": "string"
                },
                "rewardRules": {
                    "$ref": "#/definitions/api.RewardRulesReq"
                },
                "targetMarket": {
                    "type": "string"
                },
                "targetUserSegment": {
                    "type": "string"
                }
            }
        },
...
        "data.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}